feat(identity): add Action
An Action is an user intent that requires approvals from other users
before it can be performed.

In the near future Actions will be managed by a dedicated SDK module,
but until then I'm embedding them directly into the `identity` module
where they are used.

The actions we want to support are "add owner" and "remove owner".