feat(shield): allow dots (.) in identifier names
In this version of shield we'll treat dots as part of the identifier
(i.e. we won't dig much into building "javascript objects" equivalents).
It'll be up to the Environment resolving the identifier to do whatever
it prefers.

E.g. `warden.space.owners` will be treated as a single name. The
resolution env from x/warden can split by '.' and resolve 'space'
properties individually.