fix(wardenjs): set `exports` in package.json for full ESM compatibility
The "exports" key acts like a mapping from import paths to the actual
code.

With this change, the following will work in Node.js:

  import * as Foo from "@wardenprotocol/wardenjs/codegen/foo"

That will be resolved to this real path:

  node_modules/@wardenprotocol/wardenjs/dist/codegen/foo.js

It's important to note that `.js` is *mandatory* in Node.js, while it's
usually not when using other bundlers (e.g. Vite/Rollup for SpaceWard).