chore(x/warden): cleanup old or unused code
I took the chance to rename "creator" into "authority" since it makes
more sense, as it will be populated with the module account address of
x/intent.

I removed broken code for registering Action handlers, which no longer
exist.

I removed the `btl` field, present in many messages, as it now it's
specified only once in MsgNewAction and will be resposability of
x/intent to handle it.