feat(wardend/vemanager): initial version
The role of the vemanager is to provide an abstraction level for Vote
Extensions, that are usually treated as []byte.

A VoteExtensionManager acts similarly to a router for Vote Extensions,
where multiple ABCI handlers can be registered, and each handler will
only see it's own view, without having to know about vemanager at all.