feat(prophet): pricepred model handler
This is a vary bare future handler for the price prediction model we
have today, and that essentially is an experiment to showcase the tech
architecture (in fact, the date for the prediction in the request is
hardcoded).

We lack some essential features such as a way of configuring the
prediction endpoint (again, hardcoded for now).

This is also our first tentative of accepting a bytes array coming from
solidity, mapping it into a json, and then doing the other way around
for the response.