feat(x/async): implement priority queues
This commit implements the fair selection algorithm described in
CometBFT:
https://docs.cometbft.com/v1.0/spec/consensus/proposer-selection that is
used for choosing the next block proposer.

In our case, each plugin has its own priority queue. The weight of the
validators in the queue is the same staking weight. All data is
persisted in the keeper, in different new collections.