feat(x/async): Task can return errors
If the Execute() function returns an error, the error message is stored
inside the Task. The "output" field of such Tasks will remain empty, but
now it's clear that the Task is no longer pending.

The fees paid by the user are refunded.

Verifiers won't consider errored Task, as there is no output to be
verified.