feat(x/sched): inject an eth tx invoking executeCallbacks()
This tx will trigger x/sched precompile, and will be present at every
block.

The precompile will be able to fetch all the pending callbacks and do
its work.

The ante handler and ProcessProposal ensures that malicious users or
validators can't add this type of transaction to the blocks manually,
but only the injected transaction will work.