fix(x/sched): process big callbacks without gas limits
If the output of a callback was too big, things like fetching it from
the callback queue could cause to finish the gas, even before arriving
at the point of evm.Call(...) (that has a separate gas limit, paid by
the callback contract).

evm.Call(...) already reverts in case of error, so we don't need to do
it manually.